Updating First Boot Package Install Generator.app to include an Apple Setup Assistant status check
Updating First Boot Package Install Generator.app to include an Apple
Setup Assistant status check. Added functionality to check to see if
/var/db/.AppleSetupDone is present when the Mac starts up. If it is
not, the script will add add /var/db/.AppleSetupDone and
/var/db/.firstboot.delete.AppleSetupDone. Once both files have been
added, the Mac will reboot.

/var/db/.firstboot.delete.AppleSetupDone is added to provide
verification that this script added /var/db/.AppleSetupDone to suppress
the Apple Setup Assistant, which in turn allows the LoginLog
application to launch and display the log over the login window.
/var/db/.firstboot.delete.AppleSetupDone and /var/db/.AppleSetupDone
are later removed by the script to avoid interfering with an otherwise
desired launch of the Apple Setup Assistant.